Fix/tidy up 2 (#31)
* Moved intermediary search functions under SearchUtils

* minor import optimised
